requirements.txt
pymongo[srv]
python-dateutil
Flask
pandas
scipy
numpy
scikit-learn
seaborn